# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API documentation",
        "contact": {
            "email": "bekanur98@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/",
            "description": "Laravel news API server"
        }
    ],
    "paths": {
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all post with pagination",
                "operationId": "newsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "source_id",
                        "in": "query",
                        "description": "Singe of few comma separated unique IDs of sources",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2, 5, 3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "https://api.newsapi.news/api/v2/news?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "https://api.newsapi.news/api/v2/news?page=12"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "https://api.newsapi.news/api/v2/news?page=4"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "https://api.newsapi.news/api/v2/news"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "https://api.newsapi.news/api/v2/news?page=3"
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 300
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1234
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Post could not be loaded"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "Authorization and accounting endpoints",
            "externalDocs": {
                "description": "Find out more about",
                "url": "http://swagger.io"
            }
        },
        {
            "name": "News",
            "description": "News endpoints",
            "externalDocs": {
                "description": "Find out more about",
                "url": "http://swagger.io"
            }
        }
    ],
    "security": [
        []
    ]
}